5. Конвертер валюты [#424244]

+++ src/ru/job4j/converter/Converter.java
@@ -2,19 +2,19 @@ package ru.job4j.converter
 ;
 public class Converter {
 
-    public static double rubleToEuro(int value) {
-        double rsl = value / 70.0;
+    public static int rubleToEuro(int value) {
+        int rsl = value / 70;
         return rsl;
     }
 
-    public static double rubleToDollar(int value) {
-        double rsl = value/60.0; /* формула перевода рублей в доллары. */
+    public static int rubleToDollar(int value) {
+        int rsl = value/60; /* формула перевода рублей в доллары. */
         return rsl;
     }
 
     public static void main(String[] args) {
-        double euro = Converter.rubleToEuro(140);
-        double dollar = Converter.rubleToDollar(140);
+        int euro = Converter.rubleToEuro(140);
+        int dollar = Converter.rubleToDollar(140);
         System.out.println("140 rubles are " + euro + " euro.");
         System.out.println("140 rubles are " + dollar + " dollar.");
     }
